+++ src/get/getAllUsers.ts
@@ -1,4 +1,4 @@
-import { users } from './testUsers'
+import { users } from '../server/usersData'
 
 export const getAllUsers = () => {
     return users
